refactor(framework): refactor package namespace
Change packages for classes which can be separated from the implementation of specific application.
- Change name of TestController to DemoController and its endpoint from /test prefix to /demo prefix

Closes #15
