+++ src/main/java/com/Application.java
@@ -1,4 +1,4 @@
-package com.example;
+package com;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
