+++ src/main/java/com/framework/email/service/decorator/AsyncMailSender.java
@@ -1,4 +1,4 @@
-package com.example.service.email.decorator;
+package com.framework.email.service.decorator;
 
 import org.springframework.mail.MailException;
 import org.springframework.mail.javamail.JavaMailSender;
