+++ src/main/resources/application-dev.template.properties
@@ -1,11 +1,14 @@
-# database - credentials
-spring.datasource.username=root
-spring.datasource.password=root
-spring.datasource.url=jdbc:mysql://localhost:3306/balances?autoReconnect=true&amp;zeroDateTimeBehavior=convertToNull
+## Database - credentials
+spring.datasource.username=db_user
+spring.datasource.password=db_pass
+spring.datasource.url=jdbc:mysql://localhost:3306/db_name?autoReconnect=true&amp;zeroDateTimeBehavior=convertToNull
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-# database - sql logging
+
 spring.jpa.show-sql=true
 
-# email
-spring.mail.username=youremail@gmail.com
+# SMTP email configuration
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.protocol=smtp
+spring.mail.username=email@gmail.com
 spring.mail.password=password
