+++ src/main/resources/application.properties
@@ -5,11 +5,11 @@ server.port=8080
 spring.data.rest.basePath=/api
 
 # base_url
-static.base_url=http://localhost:8080
+static.base_url=http://localhost:${server.port}
 
 # database
-spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.username=db_user
+spring.datasource.password=db_pass
 spring.datasource.url=jdbc:mysql://localhost:3306/db_name?autoReconnect=true&amp;zeroDateTimeBehavior=convertToNull
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.datasource.max-active=50
