+++ src/main/resources/templates/mail/example/demo_template.html
@@ -35,7 +35,7 @@
   Advanced mail template examples at https://github.com/mailchimp/email-blueprints.
 </p>
 
-<img src="../../public/images/thymeleaf-logo.png" th:src="@{${baseUrl} + '/images/thymeleaf-logo.png'}" width="100" />
+<img src="../../../public/images/thymeleaf-logo.png" th:src="@{${baseUrl} + '/images/thymeleaf-logo.png'}" width="100" />
 <p>
   Regards, <br />
   &emsp; <em>The Thymeleaf Team</em>
